requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-sse==0.4.1
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.1
sse==1.2
tzdata==2024.2
urllib3==2.2.3
